requirements.txt
streamlit
selenium
webdriver-manager
pandas
openpyxl
requests
google-auth
google-api-python-client
python-dotenv
python-decouple
cryptography